refactor StringInput
